Creating Co-Authorship Network with Different Measures for Weights

This function creates an edge list for co-authorship networks from a data frame with a list of entities and their publications. The weight of edges can be calculated following different methods. The nodes of the network could be indifferently authors, institutions or countries.

coauth_network(
  dt,
  authors,
  articles,
  method = c("full_counting", "fractional_counting", "fractional_counting_refined"),
  cosine_normalized = FALSE
)

Arguments

dt
The data frame with authors (or institutions or countries) and the list of documents they have published.

authors
The column name of the source identifiers, that is the authors (or institutions or countries).

articles
The column name of the documents identifiers.

method
Method for calculating the edges weights, to be chosen among "full_counting","fractional_counting" or "fractional_counting_refined".

cosine_normalized
Possibility to take into account the total number of articles written by two linked authors and to normalize the weight of their link using Salton's cosine.

Value

A data.table with the authors (or institutions or countries) identifier in from and to columns, with a weight column whose values depend on the method chosen. It also keeps a copy of from and to in the Source and Target columns. This is useful is you are using the tidygraph package then, where from and to values are modified when creating a graph.

Details

Weights can be calculated with:

the "full_counting" method: the linkds between authors correspond to their absolute number of collaborations.

the "fractional_counting" method which takes into account the number of authors in each article, following (Perianes-Rodriguez et al. 2016) equation: $$\sum_{k = 1}^{M} \frac{a_{ik}.a_{jk}}{n_{k}-1}$$ with M the total number of articles, \(a_{ik}.a_{jk}\) which takes 1 if author i and j have co-written the article k, and \(n_{k}\) the number of authors for article k.

the fractional_counting_refined method, inspired by (Leydesdorff and Park 2017) which is similar to fractional_counting but which is formalised in a way that allows the sum of weights to equal the number of articles in the corpus: $$\sum_{k = 1}^{M} \frac{a_{ik}.a_{jk}.2}{n_{k}.(n_{k}-1)}$$.

In addition, it is possible to take into account the total number of collaborations of two linked authors. If cosine_normalized is set to True, the weight calculated with one of the three methods above is divided by \(\sqrt{C_{i}.C_{j}}\), with \(C_{i}\) being the number of articles co-written by author i.

References

Leydesdorff L, Park HW (2017). “Full and Fractional Counting in Bibliometric Networks.” Journal of Informetrics, 11(1), 117--120. ISSN 17511577, https://linkinghub.elsevier.com/retrieve/pii/S1751157716303133.

Perianes-Rodriguez A, Waltman L, Van Eck NJ (2016). “Constructing Bibliometric Networks: A Comparison between Full and Fractional Counting.” Journal of Informetrics, 10(4), 1178--1195. https://www.sciencedirect.com/science/article/pii/S1751157716302036?casa_token=AtzjmZ-1QmYAAAAA:2mlBPZsjGUleYi9mnybHODFw2RmMh3GHvRAuMYXygRm63cQOv07M4ixbAmJXuGq71tx2ug29baTp.

Examples

library(biblionetwork)
coauth_network(Authors_stagflation,
authors = "Author",
articles = "ItemID_Ref",
method = "fractional_counting")

#>              from            to weight        Source        Target
#>  1:     GOODWIN-T      GISSER-M   1.00     GOODWIN-T      GISSER-M
#>  2:      MANKIW-G        BALL-L   1.00      MANKIW-G        BALL-L
#>  3:        MORK-K        HALL-R   2.00        MORK-K        HALL-R
#>  4:        GALI-J   BLANCHARD-O   1.00        GALI-J   BLANCHARD-O
#>  5:       TATOM-J      RASCHE-R   1.00       TATOM-J      RASCHE-R
#>  6:       SACHS-J       BRUNO-M   3.00       SACHS-J       BRUNO-M
#>  7:      WATSON-M   CECCHETTI-S   0.25      WATSON-M   CECCHETTI-S
#>  8: SCHOENHOLTZ-K   CECCHETTI-S   0.25 SCHOENHOLTZ-K   CECCHETTI-S
#>  9:      KASMAN-B   CECCHETTI-S   0.25      KASMAN-B   CECCHETTI-S
#> 10:      HOOPER-P   CECCHETTI-S   0.25      HOOPER-P   CECCHETTI-S
#> 11:      WATSON-M      HOOPER-P   0.25      WATSON-M      HOOPER-P
#> 12: SCHOENHOLTZ-K      HOOPER-P   0.25 SCHOENHOLTZ-K      HOOPER-P
#> 13:      KASMAN-B      HOOPER-P   0.25      KASMAN-B      HOOPER-P
#> 14:      WATSON-M      KASMAN-B   0.25      WATSON-M      KASMAN-B
#> 15: SCHOENHOLTZ-K      KASMAN-B   0.25 SCHOENHOLTZ-K      KASMAN-B
#> 16:      WATSON-M SCHOENHOLTZ-K   0.25      WATSON-M SCHOENHOLTZ-K
#> 17:      GORDON-R        FRYE-J   2.00      GORDON-R        FRYE-J
#> 18:     MELTZER-A     BRUNNER-K   0.50     MELTZER-A     BRUNNER-K
#> 19:   CUKIERMAN-A     BRUNNER-K   0.50   CUKIERMAN-A     BRUNNER-K
#> 20:     MELTZER-A   CUKIERMAN-A   0.50     MELTZER-A   CUKIERMAN-A
#> 21:      LAYARD-R       GRUBB-D   0.50      LAYARD-R       GRUBB-D
#> 22:     JACKMAN-R       GRUBB-D   0.50     JACKMAN-R       GRUBB-D
#> 23:      LAYARD-R     JACKMAN-R   0.50      LAYARD-R     JACKMAN-R
#> 24:    WOODFORD-M   ROTEMBERG-J   1.00    WOODFORD-M   ROTEMBERG-J
#> 25:      WATSON-M    BERNANKE-B   0.50      WATSON-M    BERNANKE-B
#> 26:     GERTLER-M    BERNANKE-B   0.50     GERTLER-M    BERNANKE-B
#> 27:      WATSON-M     GERTLER-M   0.50      WATSON-M     GERTLER-M
#> 28:  EICHENBAUM-M       CHARI-V   0.50  EICHENBAUM-M       CHARI-V
#> 29:  CHRISTIANO-L       CHARI-V   1.00  CHRISTIANO-L       CHARI-V
#> 30:  EICHENBAUM-M  CHRISTIANO-L   0.50  EICHENBAUM-M  CHRISTIANO-L
#> 31:     GERTLER-M     CLARIDA-R   0.50     GERTLER-M     CLARIDA-R
#> 32:        GALI-J     CLARIDA-R   0.50        GALI-J     CLARIDA-R
#> 33:     GERTLER-M        GALI-J   0.50     GERTLER-M        GALI-J
#> 34: HALTIWANGER-J       DAVIS-S   1.00 HALTIWANGER-J       DAVIS-S
#> 35:    WILLIAMS-J  ORPHANIDES-A   3.00    WILLIAMS-J  ORPHANIDES-A
#> 36:        GUST-C  CHRISTIANO-L   1.00        GUST-C  CHRISTIANO-L
#> 37:      KILIAN-L      BARSKY-R   2.00      KILIAN-L      BARSKY-R
#> 38:     SARGENT-T      COGLEY-T   3.00     SARGENT-T      COGLEY-T
#> 39:  CHRISTIANO-L    ALBANESI-S   0.50  CHRISTIANO-L    ALBANESI-S
#> 40:       CHARI-V    ALBANESI-S   0.50       CHARI-V    ALBANESI-S
#> 41:       ROMER-D       ROMER-C   2.00       ROMER-D       ROMER-C
#> 42:      EUSEPI-S     BULLARD-J   1.00      EUSEPI-S     BULLARD-J
#> 43:  FITZGERALD-T  CHRISTIANO-L   1.00  FITZGERALD-T  CHRISTIANO-L
#> 44:       LIPPI-F   CUKIERMAN-A   1.00       LIPPI-F   CUKIERMAN-A
#> 45:         ZHA-T     SARGENT-T   0.50         ZHA-T     SARGENT-T
#> 46:    WILLIAMS-N     SARGENT-T   0.50    WILLIAMS-N     SARGENT-T
#> 47:         ZHA-T    WILLIAMS-N   0.50         ZHA-T    WILLIAMS-N
#> 48:      FARMER-R       BEYER-A   1.00      FARMER-R       BEYER-A
#> 49:       STARK-T       LEDUC-S   0.50       STARK-T       LEDUC-S
#> 50:        SILL-K       LEDUC-S   0.50        SILL-K       LEDUC-S
#> 51:       STARK-T        SILL-K   0.50       STARK-T        SILL-K
#> 52:    MCGREGOR-R    CHAPPELL-H   1.00    MCGREGOR-R    CHAPPELL-H
#> 53:   PESAVENTO-E     HERRERA-A   1.00   PESAVENTO-E     HERRERA-A
#> 54:     TINSLEY-P     KOZICKI-S   1.00     TINSLEY-P     KOZICKI-S
#> 55:      PAPELL-D    NIKOLSKO-R   1.00      PAPELL-D    NIKOLSKO-R
#> 56:      STRAUB-R    BILBIIE-FO   1.00      STRAUB-R    BILBIIE-FO
#> 57:        RUDD-J     BLINDER-A   1.00        RUDD-J     BLINDER-A
#> 58: EICHENGREEN-B       BORDO-M   1.00 EICHENGREEN-B       BORDO-M
#> 59:      NELSON-E     DICECIO-C   1.00      NELSON-E     DICECIO-C
#> 60:        KING-R  GOODFRIEND-M   1.00        KING-R  GOODFRIEND-M
#> 61:      TAYLOR-J       LEVIN-A   1.00      TAYLOR-J       LEVIN-A
#> 62:    WHEELOCK-D       POOLE-W   0.50    WHEELOCK-D       POOLE-W
#> 63:      RASCHE-R       POOLE-W   0.50      RASCHE-R       POOLE-W
#> 64:    WHEELOCK-D      RASCHE-R   0.50    WHEELOCK-D      RASCHE-R
#>              from            to weight        Source        Target