Creating Coupling Networks at Entity Level

This function creates the edges of a network of entities from a direct citations data frame (i.e. documents citing references). Entities could be authors, affiliations, journals, etc. Consequently, Coupling links are calculated using the coupling angle measure or the coupling strength measure. Coupling links are calculated depending of the number references two authors share, taking into account the minimum number of times two authors are citing each references. For instance, if two entities share a reference in common, the first one citing it twice (in other words, citing it in two different articles), the second one three times, the function takes two as the minimum value. In addition to the features of the coupling strength measure (see coupling_strength()) or the coupling angle measure (see biblio_coupling()), it means that, if two entities share two reference in common, the fact that the first reference is cited at least four times by the two entities, whereas the second reference is cited at least only once, the first reference contributes more to the edge weight than the second reference. This use of minimum shared reference for entities coupling comes from Zhao and Strotmann (2008) .

coupling_entity(
  dt,
  source,
  ref,
  entity,
  weight_threshold = 1,
  output_in_character = FALSE,
  method = c("coupling_strength", "coupling_angle")
)

Arguments

dt
The table with citing and cited documents.

source
the column name of the source identifiers, that is the documents that are citing.

ref
the column name of the cited references identifiers.

entity
the column name of the entity (authors, journals, institutions) that are citing.

weight_threshold
Correspond to the value of the non-normalized weights of edges. The function just keeps the edges that have a non-normalized weight superior to the weight_threshold. In a large bibliographic coupling network, you can consider for instance that sharing only one reference is not sufficient/significant for two articles (above all for two entities like authors or journals) to be linked together. This parameter could also be modified to avoid creating untractable networks with too many edges.

output_in_character
If TRUE, the function ends by transforming the from and to columns in character, to make the creation of a tidygraph graph easier.

method
Choose the method you want to use for calculating the edges weights: either "coupling_strength" like in the coupling_strength() function, or "coupling_angle" like in the biblio_coupling() function.

Value

A data.table with the entity identifiers in from and to columns, with the coupling strength or coupling angle measures in another column, as well as the method used. It also keeps a copy of from and to in the Source and Target columns. This is useful is you are using the tidygraph package then, where from and to values are modified when creating a graph.

References

Zhao D, Strotmann A (2008). “Author Bibliographic Coupling: Another Approach to Citation-Based Author Knowledge Network Analysis.” Proceedings of the American Society for Information Science and Technology, 45(1), 1--10. https://asistdl.onlinelibrary.wiley.com/doi/full/10.1002/meet.2008.1450450292.

Examples

library(biblionetwork)
# merging the references data with the citing author information in Nodes_stagflation
entity_citations <- merge(Ref_stagflation,
                          Nodes_stagflation,
                          by.x = "Citing_ItemID_Ref",
                          by.y = "ItemID_Ref")

coupling_entity(entity_citations,
                source = "Citing_ItemID_Ref",
                ref = "ItemID_Ref",
                entity = "Author.y",
                method = "coupling_angle")

#>             from          to     weight     Source      Target Weighting_method
#>    1: ALBANESI-S      BALL-L 0.02429648 ALBANESI-S      BALL-L   coupling_angle
#>    2: ALBANESI-S ROTEMBERG-J 0.03045725 ALBANESI-S ROTEMBERG-J   coupling_angle
#>    3: ALBANESI-S    BARSKY-R 0.02100729 ALBANESI-S    BARSKY-R   coupling_angle
#>    4: ALBANESI-S     BEYER-A 0.03458572 ALBANESI-S     BEYER-A   coupling_angle
#>    5: ALBANESI-S  CHAPPELL-H 0.04264014 ALBANESI-S  CHAPPELL-H   coupling_angle
#>   ---                                                                          
#> 1456:    VELDE-F     YOUNG-W 0.01601282    VELDE-F     YOUNG-W   coupling_angle
#> 1457:    VELDE-F     WEISE-C 0.02282177    VELDE-F     WEISE-C   coupling_angle
#> 1458:    VELDE-F   WACHTER-M 0.01883109    VELDE-F   WACHTER-M   coupling_angle
#> 1459:    VELDE-F WEINTRAUB-S 0.12909944    VELDE-F WEINTRAUB-S   coupling_angle
#> 1460:  WACHTER-M WEINTRAUB-S 0.14586499  WACHTER-M WEINTRAUB-S   coupling_angle