Calculating the Coupling Strength Measure for Edges

This function calculates the coupling strength measure (following Vladutz and Cook 1984 and Shen et al. 2019) from a direct citation data frame. It is a refinement of biblio_coupling(): it takes into account the frequency with which a reference shared by two articles has been cited in the whole corpus. In other words, the most cited references are less important in the links between two articles, than references that have been rarely cited. To a certain extent, it is similar to the tf-idf measure.

coupling_strength(
  dt,
  source,
  ref,
  weight_threshold = 1,
  output_in_character = FALSE
)

Arguments

dt
The data frame with citing and cited documents.

source
the column name of the source identifiers, that is the documents that are citing.

ref
the column name of the references that are cited.

weight_threshold
Corresponds to the value of the non-normalized weights of edges. The function just keeps the edges that have a non-normalized weight superior to the weight_threshold. In other words, if you set the parameter to 2, the function keeps only the edges between nodes that share at least two references in common in their bibliography. In a large bibliographic coupling network, you can consider for instance that sharing only one reference is not sufficient/significant for two articles to be linked together. This parameter could also be modified to avoid creating intractable networks with too many edges.

output_in_character
If TRUE, the function ends by transforming the from and to columns in character, to make the creation of a tidygraph graph easier.

Value

A data.table with the articles identifiers in from and to columns, with the coupling strength measure in another column. It also keeps a copy of from and to in the Source and Target columns. This is useful is you are using the tidygraph package then, where from and to values are modified when creating a graph.

References

Shen S, Zhu D, Rousseau R, Su X, Wang D (2019). “A Refined Method for Computing Bibliographic Coupling Strengths.” Journal of Informetrics, 13(2), 605--615. https://linkinghub.elsevier.com/retrieve/pii/S1751157716300244.

Vladutz G, Cook J (1984). “Bibliographic Coupling and Subject Relatedness.” Proceedings of the American Society for Information Science, 21, 204--207.

Examples

library(biblionetwork)
coupling_strength(Ref_stagflation,
source = "Citing_ItemID_Ref",
ref = "ItemID_Ref")

#>             from         to      weight     Source     Target
#>    1:     214927    2207578 0.019691698     214927    2207578
#>    2:     214927    5982867 0.005331122     214927    5982867
#>    3:     214927    8456979 0.011752248     214927    8456979
#>    4:     214927   10729971 0.046511251     214927   10729971
#>    5:     214927   16008556 0.008648490     214927   16008556
#>   ---                                                        
#> 2712: 1111111161 1111111172 0.005067554 1111111161 1111111172
#> 2713: 1111111161 1111111180 0.001168603 1111111161 1111111180
#> 2714: 1111111161 1111111183 0.002580798 1111111161 1111111183
#> 2715: 1111111172 1111111180 0.003870999 1111111172 1111111180
#> 2716: 1111111182 1111111183 0.037748271 1111111182 1111111183